fix: update import paths for various moved types
