+++ src/Model/ioActions/convertJsTypes.ts
@@ -2,12 +2,12 @@ import { isDate, isString, isSafeInteger, isBuffer } from "@nerdware/ts-type-saf
 import dayjs from "dayjs";
 import { hasDefinedProperty, isConvertibleToDate } from "../../utils/index.js";
 import type { IOActions, IOAction, IODirection } from "./types.js";
-import type { SchemaSupportedTypeStringLiterals } from "../../Schema/types.js";
-import type { BaseItem } from "../../types/itemTypes.js";
+import type { SchemaSupportedTypeStringLiteral } from "../../Schema/types/index.js";
+import type { BaseItem } from "../../types/index.js";
 
 const DDB_TYPE_MAP: Partial<
   Record<
-    SchemaSupportedTypeStringLiterals,
+    SchemaSupportedTypeStringLiteral,
     Record<IODirection, (value: NonNullable<unknown>) => NonNullable<unknown>>
   >
 > = {
