--- src/Model/ioActions/index.ts
@@ -1,2 +1,0 @@
-export * from "./ioActions.js";
-export * from "./types.js";
